Currency converter implemented

+++ backend/payment_processing/src/main/java/com/devsquad/payment_processing/model/Payment.java
@@ -3,6 +3,7 @@ package com.devsquad.payment_processing.model;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.sql.Time;
 
@@ -30,7 +31,7 @@ public class Payment {
     
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be positive")
-    private Double amount;
+    private BigDecimal amount;
     
     // Optional - if null, uses default currency
     private Integer currencyId;
@@ -52,7 +53,7 @@ public class Payment {
     private Status status;
 
     public Payment(Integer paymentId, String invoiceNumber, Long senderAccountNumber, Long receiverAccountNumber,
-                   Double amount, Integer currencyId, Integer paymentModeId, Date paymentDate, Time paymentTime,
+                   BigDecimal amount, Integer currencyId, Integer paymentModeId, Date paymentDate, Time paymentTime,
                    String description, Integer scheduleId, Status status) {
         this.paymentId = paymentId;
         this.invoiceNumber = invoiceNumber;
@@ -100,11 +101,11 @@ public class Payment {
         this.receiverAccountNumber = receiverAccountNumber;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
